requirements.txt
qrcode[pil]
pyserial
requests
